chore: clean up sidebar
